BREAKING CHANGE: Move forest plot to an extension
