+++ esm/components/data_layer/index.js
@@ -7,7 +7,6 @@
 export { default as BaseDataLayer } from './base';
 export { default as annotation_track } from './annotation_track';
 export { default as arcs } from './arcs';
-export * from './forest';
 export { default as genes } from './genes';
 export * from './line';
 export * from './scatter';
