+++ examples/interval_annotations.html
@@ -11,7 +11,7 @@
    <script src="https://cdn.jsdelivr.net/npm/d3@^5.16.0" type="text/javascript"></script>
     <script src="../dist/locuszoom.app.min.js" type="text/javascript"></script>
 
-    <!-- Helpers that use the external library to add functionality to LocusZoom -->
+    <!-- Helpers that use an extension to add functionality to LocusZoom -->
     <script src="../dist/ext/lz-intervals-track.min.js" type="application/javascript"></script>
 
     <title>LocusZoom.js ~ Interval Annotations Example</title>
